出品者かどうかを判別する条件式
◎出品者かどうかを判別する条件式がdestroyアクションにも適応されているか確認しましょう。

■理由
「「ログイン状態の場合にのみ、自身が出品した商品情報を削除できること」とは、詳細ページにおける「削除」ボタンの表示・非表示に加え、コントローラー側でも条件を設けることを指す」という実装条件を満たす為です。

+++ app/controllers/items_controller.rb
@@ -37,8 +37,12 @@ class ItemsController < ApplicationController
   
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if current_user.id == item.user_id
+      item.destroy
+      redirect_to root_path
+    else
+      redirect_to root_path
+    end
   end
 
   private
